services/model-builder/requirements.txt
scikit-learn==1.3.0
pandas
fastapi
uvicorn
pydantic
python-dotenv
joblib
